requirements.txt
asgiref==3.8.1
certifi==2024.8.30
click==8.1.7
Django==5.1.1
djangorestframework==3.15.2
joblib==1.4.2
nltk==3.9.1
regex==2024.7.24
sqlparse==0.5.1
tqdm==4.66.5
